+++ src/Simpletools/Db/Mysql/Sql.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Simpletools\Db\Mysql;
+
+class Sql
+{
+    protected $_statement = '';
+
+    public function __construct($statement)
+    {
+        $this->_statement = $statement;
+    }
+
+    public function __toString()
+    {
+        return $this->_statement;
+    }
+}